[CloudProfile] Prefactor: extract defaulting into dedicated file
