+++ pkg/apis/core/v1beta1/defaults.go
@@ -46,25 +46,6 @@ func SetDefaults_SecretBinding(obj *SecretBinding) {
 	}
 }
 
-// SetDefaults_MachineType sets default values for MachineType objects.
-func SetDefaults_MachineType(obj *MachineType) {
-	if obj.Architecture == nil {
-		obj.Architecture = pointer.String(v1beta1constants.ArchitectureAMD64)
-	}
-
-	if obj.Usable == nil {
-		obj.Usable = pointer.Bool(true)
-	}
-}
-
-// SetDefaults_VolumeType sets default values for VolumeType objects.
-func SetDefaults_VolumeType(obj *VolumeType) {
-	if obj.Usable == nil {
-		trueVar := true
-		obj.Usable = &trueVar
-	}
-}
-
 // SetDefaults_Seed sets default values for Seed objects.
 func SetDefaults_Seed(obj *Seed) {
 	if obj.Spec.Settings == nil {
@@ -452,21 +433,6 @@ func SetDefaults_ControllerRegistrationDeployment(obj *ControllerRegistrationDep
 	p := ControllerDeploymentPolicyOnDemand
 	if obj.Policy == nil {
 		obj.Policy = &p
-	}
-}
-
-// SetDefaults_MachineImageVersion sets default values for MachineImageVersion objects.
-func SetDefaults_MachineImageVersion(obj *MachineImageVersion) {
-	if len(obj.CRI) == 0 {
-		obj.CRI = []CRI{
-			{
-				Name: CRINameDocker,
-			},
-		}
-	}
-
-	if len(obj.Architectures) == 0 {
-		obj.Architectures = []string{v1beta1constants.ArchitectureAMD64}
 	}
 }
